feat: ステートメントビルダをひとつにまとめた
